+++ service/src/main/java/cn/bunny/services/service/impl/EmailUsersServiceImpl.java
@@ -110,6 +110,9 @@ public class EmailUsersServiceImpl extends ServiceImpl<EmailUsersMapper, EmailUs
      */
     @Override
     public void deleteEmailUsers(List<Long> ids) {
+        // 判断数据请求是否为空
+        if (ids.isEmpty()) throw new BunnyException(ResultCodeEnum.REQUEST_IS_EMPTY);
+        
         baseMapper.deleteBatchIdsWithPhysics(ids);
     }
